requirements.txt
torch
numpy
scipy
matplotlib
tqdm
pyyaml
jupyter
ipykernel
deepwave
segyio
